# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: İhaleZeka API — Turkish public procurement intelligence platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenders:
    get:
      operationId: listTenders
      tags: [tenders]
      summary: List tenders with optional filters
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Natural language or keyword search
        - name: il
          in: query
          schema:
            type: string
          description: Province filter (il)
        - name: tur
          in: query
          schema:
            type: string
          description: Tender type filter
        - name: usul
          in: query
          schema:
            type: string
          description: Procurement method filter
        - name: minBedel
          in: query
          schema:
            type: number
          description: Minimum estimated value
        - name: maxBedel
          in: query
          schema:
            type: number
          description: Maximum estimated value
        - name: idare
          in: query
          schema:
            type: string
          description: Agency name filter
        - name: cpv
          in: query
          schema:
            type: string
          description: CPV code filter
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of tenders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenderPage"
  /tenders/{id}:
    get:
      operationId: getTender
      tags: [tenders]
      summary: Get tender details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tender detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenderDetail"
        "404":
          description: Not found

  # ─── Matches ─────────────────────────────────────────────────────────────────
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List proposals
      responses:
        "200":
          description: Proposals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"

    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create a new proposal draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"

    patch:
      operationId: updateProposal
      tags: [proposals]
      summary: Update proposal content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalUpdate"
      responses:
        "200":
          description: Updated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"

  # ─── Competitors ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Tender:
      type: object
      required: [id, ikn, title, agencyName, type, method, estimatedValue, deadline, status, il]
      properties:
        id:
          type: integer
        ikn:
          type: string
          description: İhale Kayıt Numarası
        title:
          type: string
        agencyName:
          type: string
        agencyLogoUrl:
          type: ["string", "null"]
        type:
          type: string
          description: Tender type (e.g. mal alımı, hizmet alımı, yapım işi)
        method:
          type: string
          description: Procurement method (e.g. açık ihale usulü)
        estimatedValue:
          type: number
        deadline:
          type: string
          format: date-time
        cpvCodes:
          type: array
          items:
            type: string
        il:
          type: string
          description: Province
        status:
          type: string
          enum: [active, cancelled, awarded, draft]
        sourceSystem:
          type: string
          default: ekap
    TenderDetail:
      allOf:
        - $ref: "#/components/schemas/Tender"
        - type: object
          properties:
            description:
              type: ["string", "null"]
            qualificationCriteria:
              type: array
              items:
                type: string
            documentsRequired:
              type: array
              items:
                type: string
            rawDocsUrls:
              type: array
              items:
                type: string
    TenderPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Tender"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── Match ──────────────────────────────────────────────────────────────────
...
    Proposal:
      type: object
      required: [id, tenderId, status, createdAt]
      properties:
        id:
          type: integer
        tenderId:
          type: integer
        tenderTitle:
          type: ["string", "null"]
        contentJson:
          type: ["string", "null"]
          description: JSON-encoded proposal sections
        status:
          type: string
          enum: [draft, finalized, submitted]
        version:
          type: integer
          default: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProposalInput:
      type: object
      required: [tenderId]
      properties:
        tenderId:
          type: integer
        contentJson:
          type: string
    ProposalUpdate:
      type: object
      properties:
        contentJson:
          type: string
        status:
          type: string
          enum: [draft, finalized, submitted]

    # ─── Competitors ─────────────────────────────────────────────────────────────
...